requirements.txt
duckdb
pandas
pyyaml
rapidfuzz
pydantic
google-generativeai
fastapi
uvicorn
httpx
pinecone-client
sentence-transformers
